landscape/old/requirements.txt
fastapi==0.110.2
hdbscan==0.8.33
matplotlib==3.8.4
nltk==3.7
numpy==1.24.3
pandas==2.2.2
plotly==5.21.0
psycopg2==2.9.9
pydantic==1.10.12
python-dotenv==1.0.1
Requests==2.31.0
retry==0.9.2
scikit_learn==1.2.1
scipy==1.13.0
sshtunnel==0.4.0
umap_learn==0.5.3
